+++ cmd/migrate.go
@@ -7,7 +7,7 @@
 package cmd
 
 import (
-	"github.com/jkaninda/pg-bkup/pkg"
+	"github.com/jkaninda/pg-bkup/internal"
 	"github.com/jkaninda/pg-bkup/utils"
 	"github.com/spf13/cobra"
 )
@@ -17,7 +17,7 @@ var MigrateCmd = &cobra.Command{
 	Short: "Migrate database from a source database to a target database",
 	Run: func(cmd *cobra.Command, args []string) {
 		if len(args) == 0 {
-			pkg.StartMigration(cmd)
+			internal.StartMigration(cmd)
 		} else {
 			utils.Fatal(`"migrate" accepts no argument %q`, args)
