+++ cmd/restore.go
@@ -7,7 +7,7 @@
 package cmd
 
 import (
-	"github.com/jkaninda/pg-bkup/pkg"
+	"github.com/jkaninda/pg-bkup/internal"
 	"github.com/jkaninda/pg-bkup/utils"
 	"github.com/spf13/cobra"
 )
@@ -18,7 +18,7 @@ var RestoreCmd = &cobra.Command{
 	Example: utils.RestoreExample,
 	Run: func(cmd *cobra.Command, args []string) {
 		if len(args) == 0 {
-			pkg.StartRestore(cmd)
+			internal.StartRestore(cmd)
 		} else {
 			utils.Fatal(`"restore" accepts no argument %q`, args)
